learn_with_me/pyproject.toml
[tool.poetry]
name = "learn-with-me"
version = "0.0.1"
description = ""
authors = ["Davi Nascimento de Paula <davi.paula@gmail.com>"]
license = "MIT 2.0"

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.52.0"
webvtt-py = "^0.4.5"
google-api-python-client = "^1.12.5"
google-auth-oauthlib = "^0.4.2"
google-auth-httplib2 = "^0.0.4"
pytest = "^6.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
